+++ internal/constants/errors.go
@@ -570,6 +570,7 @@ var (
 	ErrBootstrapAuthFailed         = errors.New("bootstrap: authentication failed")
 	ErrBootstrapLoopbackOnly       = errors.New("CSR auto-issue only available over loopback")
 	ErrBootstrapUserDisabled       = errors.New("bootstrap user is disabled, cannot rotate")
+	ErrBootstrapUserDisabledEnroll = errors.New("bootstrap user is disabled, cannot enroll")
 	ErrBootstrapInitialSetupOnly   = errors.New("bootstrap only available for initial setup")
 	ErrCLIEnrollmentLoopbackOnly   = errors.New("CLI enrollment only available over loopback")
 	ErrCLIEnrollmentAfterBootstrap = errors.New("CLI enrollment only available after bootstrap")
@@ -933,31 +934,31 @@ var (
 	ErrBlobStoreParseTime     = errors.New("failed to parse blob creation time")
 
 	// Auth service errors
-	ErrIdentityDisabled            = errors.New("identity disabled")
-	ErrOperatorIdentityDisabled    = errors.New("operator identity disabled")
-	ErrOperatorSessionExpired      = errors.New("operator session expired")
-	ErrMTLSCertRequired            = errors.New("mTLS client certificate required")
-	ErrMTLSCertRevoked             = errors.New("mTLS client certificate revoked or invalid")
-	ErrProtocolAuthRequired        = errors.New("protocol authentication required")
-	ErrMTLSIdentityMismatch        = errors.New("mTLS identity mismatch")
-	ErrSessionLoadFailed           = errors.New("failed to load session")
-	ErrSessionParseFailed          = errors.New("failed to parse session")
-	ErrCLISessionExpired           = errors.New("CLI session expired")
-	ErrIdentityValidationFailed    = errors.New("identity validation failed")
-	ErrAppPolicyNotFound           = errors.New("app policy not found")
-	ErrInvalidAppPolicy            = errors.New("invalid app policy")
-	ErrPrivilegedEndpointAccess    = errors.New("external apps cannot access privileged endpoints")
-	ErrWebSessionCookieRequired    = errors.New("web session cookie required")
-	ErrWebSessionCookieInvalid     = errors.New("invalid web session cookie")
-	ErrWebSessionValidationFailed  = errors.New("web session validation failed")
-	ErrWebSessionNotFound          = errors.New("web session not found")
-	ErrWebSessionExpired           = errors.New("web session expired")
-	ErrJWTAuthNotConfigured        = errors.New("JWT authentication not configured")
-	ErrJWTTokenMissing             = errors.New("missing JWT token")
-	ErrJWTSignatureMissing         = errors.New("missing JWT bearer token")
-	ErrJWTSessionSubjectMissing    = errors.New("JWT missing subject claim")
-	ErrJWTInvalidToken             = errors.New("invalid JWT token")
-	ErrUserCreationFailed          = errors.New("user creation failed")
-	ErrCLISessionInvalid           = errors.New("invalid CLI session")
-	ErrCLICertBindingCheckFailed   = errors.New("CLI cert binding check failed")
+	ErrIdentityDisabled           = errors.New("identity disabled")
+	ErrOperatorIdentityDisabled   = errors.New("operator identity disabled")
+	ErrOperatorSessionExpired     = errors.New("operator session expired")
+	ErrMTLSCertRequired           = errors.New("mTLS client certificate required")
+	ErrMTLSCertRevoked            = errors.New("mTLS client certificate revoked or invalid")
+	ErrProtocolAuthRequired       = errors.New("protocol authentication required")
+	ErrMTLSIdentityMismatch       = errors.New("mTLS identity mismatch")
+	ErrSessionLoadFailed          = errors.New("failed to load session")
+	ErrSessionParseFailed         = errors.New("failed to parse session")
+	ErrCLISessionExpired          = errors.New("CLI session expired")
+	ErrIdentityValidationFailed   = errors.New("identity validation failed")
+	ErrAppPolicyNotFound          = errors.New("app policy not found")
+	ErrInvalidAppPolicy           = errors.New("invalid app policy")
+	ErrPrivilegedEndpointAccess   = errors.New("external apps cannot access privileged endpoints")
+	ErrWebSessionCookieRequired   = errors.New("web session cookie required")
+	ErrWebSessionCookieInvalid    = errors.New("invalid web session cookie")
+	ErrWebSessionValidationFailed = errors.New("web session validation failed")
+	ErrWebSessionNotFound         = errors.New("web session not found")
+	ErrWebSessionExpired          = errors.New("web session expired")
+	ErrJWTAuthNotConfigured       = errors.New("JWT authentication not configured")
+	ErrJWTTokenMissing            = errors.New("missing JWT token")
+	ErrJWTSignatureMissing        = errors.New("missing JWT bearer token")
+	ErrJWTSessionSubjectMissing   = errors.New("JWT missing subject claim")
+	ErrJWTInvalidToken            = errors.New("invalid JWT token")
+	ErrUserCreationFailed         = errors.New("user creation failed")
+	ErrCLISessionInvalid          = errors.New("invalid CLI session")
+	ErrCLICertBindingCheckFailed  = errors.New("CLI cert binding check failed")
 )
