+++ test/agentic_tool_emulator/scenarios/mcp_a2a_test.go
@@ -175,20 +175,13 @@ func TestFirstToolEdgeCases(t *testing.T) {
 func TestMCPScenarios(t *testing.T) {
 	scenarios := mcpScenarios()
 
-	// Should have 3 MCP scenarios
-	if len(scenarios) != 3 {
-		t.Errorf("mcpScenarios should return 3 scenarios, got %d", len(scenarios))
-	}
-
-	// All should require Doctrine posture
-	for _, sc := range scenarios {
-		if sc.RequiresPosture != Doctrine {
-			t.Errorf("MCP scenario %q should require Doctrine posture, got %q", sc.Name, sc.RequiresPosture)
-		}
+	// Should have 5 MCP scenarios (3 core + 2 healthcare)
+	if len(scenarios) != 5 {
+		t.Errorf("mcpScenarios should return 5 scenarios, got %d", len(scenarios))
 	}
 
 	// Should have expected names
-	expectedNames := []string{"mcp-plain", "mcp-advanced", "mcp-secured"}
+	expectedNames := []string{"mcp-plain", "mcp-advanced", "mcp-secured", "healthcare-success", "healthcare-phi-blocked"}
 	nameSet := make(map[string]bool)
 	for _, sc := range scenarios {
 		nameSet[sc.Name] = true
@@ -232,9 +225,11 @@ func TestMCPScenarioNames(t *testing.T) {
 	scenarios := mcpScenarios()
 
 	expectedNames := map[string]bool{
-		"mcp-plain":    true,
-		"mcp-advanced": true,
-		"mcp-secured":  true,
+		"mcp-plain":             true,
+		"mcp-advanced":          true,
+		"mcp-secured":           true,
+		"healthcare-success":    true,
+		"healthcare-phi-blocked": true,
 	}
 
 	for _, sc := range scenarios {
@@ -264,9 +259,11 @@ func TestMCPScenarioPersonas(t *testing.T) {
 	scenarios := mcpScenarios()
 
 	expectedPersonas := map[string]string{
-		"mcp-plain":    "claude-desktop",
-		"mcp-advanced": "cursor",
-		"mcp-secured":  "enterprise-agent",
+		"mcp-plain":              "claude-desktop",
+		"mcp-advanced":           "cursor",
+		"mcp-secured":            "enterprise-agent",
+		"healthcare-success":     "clinical-agent",
+		"healthcare-phi-blocked": "clinical-agent",
 	}
 
 	for _, sc := range scenarios {
@@ -306,9 +303,11 @@ func TestMCPScenarioTitles(t *testing.T) {
 	scenarios := mcpScenarios()
 
 	expectedTitles := map[string]string{
-		"mcp-plain":    "Plain MCP tool call",
-		"mcp-advanced": "Advanced MCP: resources, prompts, chained calls",
-		"mcp-secured":  "MCP with simple security (mTLS/API key + L1 gate)",
+		"mcp-plain":              "Plain MCP tool call",
+		"mcp-advanced":           "Advanced MCP: resources, prompts, chained calls",
+		"mcp-secured":            "MCP with simple security (mTLS/API key + L1 gate)",
+		"healthcare-success":     "Authorized FHIR PA Submission",
+		"healthcare-phi-blocked": "PHI Exfiltration Blocked by Doctrine",
 	}
 
 	for _, sc := range scenarios {
